+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: `SlideModal`: Add new component.
 - Fix: `CloseButton`: Use `CrossIconNext` icon.
 - Feature: `Field`: Add `tag` prop.
+- Feature: `Field`: Add `center` prop to `Field.Label`.
+- Fix: `TextInputWithButton`: Fix mobile styles.
 
 ## [11.7.1] - 2022-07-07